Prometheus: Move parsing of query and response to separate files
