+++ pkg/tsdb/prometheus/response_parser.go
@@ -1,0 +1,276 @@
+package prometheus
+
+import (
+	"fmt"
+	"math"
+	"sort"
+	"strings"
+	"time"
+
+	"github.com/grafana/grafana-plugin-sdk-go/data"
+	apiv1 "github.com/prometheus/client_golang/api/prometheus/v1"
+	"github.com/prometheus/common/model"
+)
+
+func parseResponse(value map[PrometheusQueryType]interface{}, query *PrometheusQuery) (data.Frames, error) {
+	frames := data.Frames{}
+
+	for _, value := range value {
+		matrix, ok := value.(model.Matrix)
+		if ok {
+			matrixFrames := matrixToDataFrames(matrix, query)
+			frames = append(frames, matrixFrames...)
+			continue
+		}
+
+		vector, ok := value.(model.Vector)
+		if ok {
+			vectorFrames := vectorToDataFrames(vector, query)
+			frames = append(frames, vectorFrames...)
+			continue
+		}
+
+		scalar, ok := value.(*model.Scalar)
+		if ok {
+			scalarFrames := scalarToDataFrames(scalar, query)
+			frames = append(frames, scalarFrames...)
+			continue
+		}
+
+		exemplar, ok := value.([]apiv1.ExemplarQueryResult)
+		if ok {
+			exemplarFrames := exemplarToDataFrames(exemplar, query)
+			frames = append(frames, exemplarFrames...)
+			continue
+		}
+	}
+	return frames, nil
+}
+
+func matrixToDataFrames(matrix model.Matrix, query *PrometheusQuery) data.Frames {
+	frames := data.Frames{}
+
+	for _, v := range matrix {
+		tags := make(map[string]string, len(v.Metric))
+		timeVector := make([]time.Time, 0, len(v.Values))
+		values := make([]float64, 0, len(v.Values))
+		for k, v := range v.Metric {
+			tags[string(k)] = string(v)
+		}
+		for _, k := range v.Values {
+			timeVector = append(timeVector, time.Unix(k.Timestamp.Unix(), 0).UTC())
+			values = append(values, float64(k.Value))
+		}
+		name := formatLegend(v.Metric, query)
+		frame := data.NewFrame(name,
+			data.NewField("Time", nil, timeVector),
+			data.NewField("Value", tags, values).SetConfig(&data.FieldConfig{DisplayNameFromDS: name}))
+		frame.Meta = &data.FrameMeta{
+			Custom: map[string]string{
+				"resultType": "matrix",
+			},
+		}
+		frames = append(frames, frame)
+	}
+	return frames
+}
+
+func scalarToDataFrames(scalar *model.Scalar, query *PrometheusQuery) data.Frames {
+	timeVector := []time.Time{time.Unix(scalar.Timestamp.Unix(), 0).UTC()}
+	values := []float64{float64(scalar.Value)}
+	name := fmt.Sprintf("%g", values[0])
+	frame := data.NewFrame(name,
+		data.NewField("Time", nil, timeVector),
+		data.NewField("Value", nil, values).SetConfig(&data.FieldConfig{DisplayNameFromDS: name}))
+	frame.Meta = &data.FrameMeta{
+		Custom: map[string]string{
+			"resultType": "scalar",
+		},
+	}
+
+	frames := data.Frames{frame}
+	return frames
+}
+
+func vectorToDataFrames(vector model.Vector, query *PrometheusQuery) data.Frames {
+	frames := data.Frames{}
+	for _, v := range vector {
+		name := formatLegend(v.Metric, query)
+		tags := make(map[string]string, len(v.Metric))
+		timeVector := []time.Time{time.Unix(v.Timestamp.Unix(), 0).UTC()}
+		values := []float64{float64(v.Value)}
+		for k, v := range v.Metric {
+			tags[string(k)] = string(v)
+		}
+		frame := data.NewFrame(name,
+			data.NewField("Time", nil, timeVector),
+			data.NewField("Value", tags, values).SetConfig(&data.FieldConfig{DisplayNameFromDS: name}))
+		frame.Meta = &data.FrameMeta{
+			Custom: map[string]string{
+				"resultType": "vector",
+			},
+		}
+		frames = append(frames, frame)
+	}
+
+	return frames
+}
+
+func exemplarToDataFrames(response []apiv1.ExemplarQueryResult, query *PrometheusQuery) data.Frames {
+	frames := data.Frames{}
+	events := make([]ExemplarEvent, 0)
+	for _, exemplarData := range response {
+		for _, exemplar := range exemplarData.Exemplars {
+			event := ExemplarEvent{}
+			exemplarTime := time.Unix(exemplar.Timestamp.Unix(), 0).UTC()
+			event.Time = exemplarTime
+			event.Value = float64(exemplar.Value)
+			event.Labels = make(map[string]string)
+			for label, value := range exemplar.Labels {
+				event.Labels[string(label)] = string(value)
+			}
+			for seriesLabel, seriesValue := range exemplarData.SeriesLabels {
+				event.Labels[string(seriesLabel)] = string(seriesValue)
+			}
+
+			events = append(events, event)
+		}
+	}
+
+	//Sampling of exemplars
+	bucketedExemplars := make(map[string][]ExemplarEvent)
+	values := make([]float64, 0)
+
+	//Create bucketed exemplars based on aligned timestamp
+	for _, event := range events {
+		alignedTs := fmt.Sprintf("%.0f", math.Floor(float64(event.Time.Unix())/query.Step.Seconds())*query.Step.Seconds())
+		_, ok := bucketedExemplars[alignedTs]
+		if !ok {
+			bucketedExemplars[alignedTs] = make([]ExemplarEvent, 0)
+		}
+
+		bucketedExemplars[alignedTs] = append(bucketedExemplars[alignedTs], event)
+		values = append(values, event.Value)
+	}
+
+	//Calculate standard deviation
+	standardDeviation := deviation(values)
+
+	//Create slice with all of the bucketed exemplars
+	sampledBuckets := make([]string, len(bucketedExemplars))
+	for bucketTimes := range bucketedExemplars {
+		sampledBuckets = append(sampledBuckets, bucketTimes)
+	}
+	sort.Strings(sampledBuckets)
+
+	//Sample exemplars based ona value, so we are not showing too many of them
+	sampleExemplars := make([]ExemplarEvent, 0)
+	for _, bucket := range sampledBuckets {
+		exemplarsInBucket := bucketedExemplars[bucket]
+		if len(exemplarsInBucket) == 1 {
+			sampleExemplars = append(sampleExemplars, exemplarsInBucket[0])
+		} else {
+			bucketValues := make([]float64, len(exemplarsInBucket))
+			for _, exemplar := range exemplarsInBucket {
+				bucketValues = append(bucketValues, exemplar.Value)
+			}
+			sort.Slice(bucketValues, func(i, j int) bool {
+				return bucketValues[i] > bucketValues[j]
+			})
+
+			sampledBucketValues := make([]float64, 0)
+			for _, value := range bucketValues {
+				if len(sampledBucketValues) == 0 {
+					sampledBucketValues = append(sampledBucketValues, value)
+				} else {
+					// Then take values only when at least 2 standard deviation distance to previously taken value
+					prev := sampledBucketValues[len(sampledBucketValues)-1]
+					if standardDeviation != 0 && prev-value >= float64(2)*standardDeviation {
+						sampledBucketValues = append(sampledBucketValues, value)
+					}
+				}
+			}
+			for _, valueBucket := range sampledBucketValues {
+				for _, exemplar := range exemplarsInBucket {
+					if exemplar.Value == valueBucket {
+						sampleExemplars = append(sampleExemplars, exemplar)
+					}
+				}
+			}
+		}
+	}
+
+	// Create DF from sampled exemplars
+	timeVector := make([]time.Time, 0, len(sampleExemplars))
+	valuesVector := make([]float64, 0, len(sampleExemplars))
+	labelsVector := make(map[string][]string, len(sampleExemplars))
+
+	for _, exemplar := range sampleExemplars {
+		timeVector = append(timeVector, exemplar.Time)
+		valuesVector = append(valuesVector, exemplar.Value)
+
+		for label, value := range exemplar.Labels {
+			if labelsVector[label] == nil {
+				labelsVector[label] = make([]string, 0)
+			}
+
+			labelsVector[label] = append(labelsVector[label], value)
+		}
+	}
+
+	frame := data.NewFrame("exemplar",
+		data.NewField("Time", nil, timeVector),
+		data.NewField("Value", nil, valuesVector))
+
+	for label, vector := range labelsVector {
+		frame.Fields = append(frame.Fields, data.NewField(label, nil, vector))
+	}
+
+	frame.Meta = &data.FrameMeta{
+		Custom: map[string]PrometheusQueryType{
+			"resultType": "exemplar",
+		},
+	}
+
+	frames = append(frames, frame)
+	return frames
+}
+
+func deviation(values []float64) float64 {
+	var sum, mean, sd float64
+	valuesLen := float64(len(values))
+	for _, value := range values {
+		sum += value
+	}
+	mean = sum / valuesLen
+	for j := 0; j < len(values); j++ {
+		sd += math.Pow(values[j]-mean, 2)
+	}
+	return math.Sqrt(sd / (valuesLen - 1))
+}
+
+func formatLegend(metric model.Metric, query *PrometheusQuery) string {
+	var legend string
+
+	if query.LegendFormat == "" {
+		legend = metric.String()
+	} else {
+		result := legendFormat.ReplaceAllFunc([]byte(query.LegendFormat), func(in []byte) []byte {
+			labelName := strings.Replace(string(in), "{{", "", 1)
+			labelName = strings.Replace(labelName, "}}", "", 1)
+			labelName = strings.TrimSpace(labelName)
+			if val, exists := metric[model.LabelName(labelName)]; exists {
+				return []byte(val)
+			}
+			return []byte{}
+		})
+		legend = string(result)
+	}
+
+	// If legend is empty brackets, use query expression
+	if legend == "{}" {
+		legend = query.Expr
+	}
+
+	return legend
+}
